requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
requests==2.31.0
python-dotenv==1.0.0
python-socketio==5.11.0
eventlet==0.33.3
